Bean Picking Created, Limits on Additives
Bean Picking is now finished™. It gets and displays the stats of each bean and how you can use them so that's neat! The stats also work correctly too so that's also neat! The issue right now is restarting the minigame. Uhhhh Idk how to do that. So i'll yell at eric tomorrow over this exisential crisis.

Limits added on additives, originally i said 3 pumps but lets just go with 5 for now I guess. hehe HooHoo. This number is easily adjustable so it's not too big of a deal and obviously we're playtesting to see how we can affect it anyways so it's FINE