Bean Reset implemented + UI Font Changes & More
Lots of changes here, probably should've pushed earlier but oh whell

Bean Resetting is implemented, had to fiddle with the code but you can remove beans from your inventory by clicking them now so that is easier, the scene will reset itself so maybe do something about the music? It'll reboot hte music which is an issue and a polish thing so look into that probably. For now, everything functions pretty well and will do what is needed so good job niko

- UI Font has been updated in Cooking, everywhere else not yet. I'll handle that tomorrow during hte main merge so that will be fun and not excruciating.

- Book is updated on the cooking thing but I miss the bookmarks, might ask alicia to get a new version of that not sure yet but otherwise, buttons have pretty wow light up buttons when you hover over stuff now so neato botito idk man